#179 extensions now initialised in ShellViewModel on startup
